+++ jexxa-adapter-api/src/main/java/io/jexxa/adapterapi/invocation/transaction/TransactionHandler.java
@@ -1,0 +1,8 @@
+package io.jexxa.adapterapi.invocation.transaction;
+
+public interface TransactionHandler {
+    void initTransaction();
+    void closeTransaction();
+
+    void rollback();
+}