+++ jexxa-core/src/main/java/io/jexxa/core/JexxaMain.java
@@ -1,5 +1,6 @@
 package io.jexxa.core;
 
+import io.jexxa.adapterapi.JexxaContext;
 import io.jexxa.adapterapi.drivingadapter.HealthCheck;
 import io.jexxa.adapterapi.drivingadapter.IDrivingAdapter;
 import io.jexxa.core.convention.AdapterConvention;
@@ -86,7 +87,7 @@ public final class JexxaMain
     {
         Objects.requireNonNull(applicationProperties);
         Objects.requireNonNull(context);
-        JexxaBanner.clear();
+        JexxaContext.init();
 
         // Handle properties in following forder:
         // 0. Add default JEXXA_CONTEXT_MAIN
@@ -346,7 +347,7 @@ public final class JexxaMain
             JexxaLogger.getLogger(JexxaMain.class).info("BoundedContext '{}' successfully stopped", getBoundedContext().contextName());
         }
         executorService.shutdown();
-        JexxaBanner.clear();
+        JexxaContext.cleanup();
     }
 
     @SuppressWarnings("UnusedReturnValue")