+++ jexxa-core/src/main/java/io/jexxa/infrastructure/drivenadapterstrategy/persistence/jdbc/JDBCConnection.java
@@ -19,13 +19,34 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Properties;
 
+
 public class JDBCConnection implements AutoCloseable
 {
+
+    public enum IsolationLevel {
+        READ_UNCOMMITTED(Connection.TRANSACTION_READ_UNCOMMITTED),
+        READ_COMMITTED(Connection.TRANSACTION_READ_COMMITTED),
+        REPEATABLE_READ(Connection.TRANSACTION_REPEATABLE_READ),
+        SERIALIZABLE(Connection.TRANSACTION_SERIALIZABLE);
+        private final int transactionID;
+        IsolationLevel(int transactionID)
+        {
+            this.transactionID = transactionID;
+        }
+
+        public int getID()
+        {
+            return transactionID;
+        }
+    }
+
 
     public static final int NO_TIMEOUT = 0;
 
     private Connection connection;
     private final Properties properties;
+    private IsolationLevel isolationLevel;
+    private boolean autoCommit = true;
 
     private static final Logger LOGGER = JexxaLogger.getLogger(JDBCConnection.class);
 
@@ -37,9 +58,45 @@ public class JDBCConnection implements AutoCloseable
 
         autocreateDatabase(properties);
 
-        this.connection = initJDBCConnection(properties);
+        this.connection = initJDBCConnection(properties, autoCommit);
         this.properties = properties;
+        setIsolationLevel();
     }
+
+    @SuppressWarnings("MagicConstant") // The enum IsolationLevel ensures correct constant
+    public void setIsolationLevel(IsolationLevel isolationLevel) {
+        this.isolationLevel = isolationLevel;
+        if (isolationLevel != null)
+        {
+            try {
+                if ( getConnection().getMetaData().supportsTransactionIsolationLevel(isolationLevel.getID()))
+                {
+                    getConnection().setTransactionIsolation(isolationLevel.getID());
+                }
+            } catch (SQLException e)
+            {
+                throw new IllegalStateException("Configuring IsolationLevel failed. Check if isolation level is supported by your DB", e);
+            }
+        }
+    }
+
+    protected void setIsolationLevel()
+    {
+        if (isolationLevel == null
+                && properties.containsKey(JexxaJDBCProperties.JEXXA_JDBC_TRANSACTION_ISOLATION_LEVEL))
+        {
+            isolationLevel = IsolationLevel.valueOf(properties
+                    .getProperty(JexxaJDBCProperties.JEXXA_JDBC_TRANSACTION_ISOLATION_LEVEL)
+                    .toUpperCase(Locale.ROOT).replace('-', '_')
+            );
+        }
+
+        if (isolationLevel != null)
+        {
+            setIsolationLevel(isolationLevel);
+        }
+    }
+
 
     public final void autocreateDatabase(final Properties properties)
     {
@@ -91,6 +148,48 @@ public class JDBCConnection implements AutoCloseable
         }
 
         return this;
+    }
+
+    public void enableAutoCommit()
+    {
+        autoCommit = true;
+        try {
+            connection.setAutoCommit(true);
+        } catch (SQLException e) {
+            throw new IllegalStateException("Could not enable auto commit for JDBC connection", e);
+        }
+    }
+
+    public void disableAutoCommit()
+    {
+        autoCommit = false;
+        try {
+            connection.setAutoCommit(false);
+        } catch (SQLException e) {
+            throw new IllegalStateException("Could not disable auto commit for JDBC connection", e);
+        }
+    }
+
+    public void commit()
+    {
+        if (!autoCommit) {
+            try {
+                connection.commit();
+            } catch (SQLException e) {
+                throw new IllegalStateException("Could not perform commit operation on JDBC connection", e);
+            }
+        }
+    }
+
+    public void rollback()
+    {
+        if (!autoCommit) {
+            try {
+                connection.rollback();
+            } catch (SQLException e) {
+                throw new IllegalStateException("Could not perform rollback operation on JDBC connection", e);
+            }
+        }
     }
 
     @SuppressWarnings("java:S1172")
@@ -179,12 +278,13 @@ public class JDBCConnection implements AutoCloseable
     {
         if ( connection == null )
         {
-            connection = initJDBCConnection(properties);
+            connection = initJDBCConnection(properties, autoCommit);
+            setIsolationLevel();
         }
         return connection;
     }
 
-    private static Connection initJDBCConnection(Properties properties)
+    private static Connection initJDBCConnection(Properties properties, boolean autoCommit)
     {
         var username = new Secret(properties, JexxaJDBCProperties.JEXXA_JDBC_USERNAME, JexxaJDBCProperties.JEXXA_JDBC_FILE_USERNAME);
         var password = new Secret(properties, JexxaJDBCProperties.JEXXA_JDBC_PASSWORD, JexxaJDBCProperties.JEXXA_JDBC_FILE_PASSWORD);
@@ -196,7 +296,7 @@ public class JDBCConnection implements AutoCloseable
                     password.getSecret()
             );
 
-            connection.setAutoCommit(true);
+            connection.setAutoCommit(autoCommit);
             return connection;
         }
         catch (SQLException e)