+++ jexxa-core/src/main/java/io/jexxa/infrastructure/drivenadapterstrategy/persistence/repository/jdbc/JDBCRepository.java
@@ -1,26 +1,22 @@
 package io.jexxa.infrastructure.drivenadapterstrategy.persistence.repository.jdbc;
 
+import io.jexxa.adapterapi.invocation.transaction.TransactionHandler;
+import io.jexxa.adapterapi.invocation.transaction.TransactionManager;
 import io.jexxa.infrastructure.drivenadapterstrategy.persistence.jdbc.JDBCConnection;
-import io.jexxa.utils.function.ThrowingConsumer;
+import io.jexxa.infrastructure.drivenadapterstrategy.persistence.jdbc.JDBCConnectionPool;
+import io.jexxa.utils.JexxaLogger;
 
 import java.util.Objects;
 import java.util.Properties;
 
-public abstract class JDBCRepository  implements AutoCloseable
-{
-    private final JDBCConnection jdbcConnection;
+public abstract class JDBCRepository implements TransactionHandler {
+    private final Properties properties;
 
     protected JDBCRepository(Properties properties)
     {
-        Objects.requireNonNull(properties);
-
-        this.jdbcConnection = new JDBCConnection(properties);
-    }
-
-    @SuppressWarnings("ResultOfMethodCallIgnored")
-    public void close()
-    {
-        ThrowingConsumer.exceptionLogger(JDBCConnection::close);
+        this.properties = Objects.requireNonNull(properties);
+        getConnection(); // To ensure that connection is valid
+        TransactionManager.registerTransactionHandler(this);
     }
 
     /**
@@ -29,9 +25,27 @@ public abstract class JDBCRepository  implements AutoCloseable
      * @throws IllegalStateException if JDBCConnection can not be reset
      * @return JDBCConnection that is in a valid state.
      */
-    public final JDBCConnection getConnection()
+    public JDBCConnection getConnection()
     {
-        return jdbcConnection.validateConnection();
+        return JDBCConnectionPool.getConnection(properties, this);
+    }
+    public void initTransaction()
+    {
+        getConnection().disableAutoCommit();
+    }
+    public void closeTransaction()
+    {
+        getConnection().commit();
+        getConnection().enableAutoCommit();
     }
 
+    public void rollback()
+    {
+        try {
+            getConnection().rollback();
+        } catch (IllegalStateException e)
+        {
+            JexxaLogger.getLogger(getClass()).error("An exception occurred during rollback. Reason: {}", e.getMessage());
+        }
+    }
 }