+++ jexxa-core/src/test/java/io/jexxa/infrastructure/drivenadapterstrategy/messaging/jms/JMSSenderIT.java
@@ -7,6 +7,7 @@ import io.jexxa.application.domain.model.JexxaValueObject;
 import io.jexxa.core.JexxaMain;
 import io.jexxa.infrastructure.drivenadapterstrategy.messaging.MessageSender;
 import io.jexxa.infrastructure.drivenadapterstrategy.messaging.MessageSenderManager;
+import io.jexxa.infrastructure.drivenadapterstrategy.outbox.TransactionalOutboxSender;
 import io.jexxa.infrastructure.drivingadapter.messaging.JMSAdapter;
 import io.jexxa.infrastructure.utils.messaging.QueueListener;
 import io.jexxa.infrastructure.utils.messaging.TopicListener;
@@ -16,6 +17,8 @@ import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.parallel.Execution;
 import org.junit.jupiter.api.parallel.ExecutionMode;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 
 import javax.jms.BytesMessage;
 import javax.jms.Connection;
@@ -24,7 +27,7 @@ import javax.jms.TextMessage;
 import java.time.Duration;
 import java.util.Properties;
 import java.util.concurrent.TimeUnit;
-
+import java.util.stream.Stream;
 
 import static io.jexxa.infrastructure.utils.messaging.QueueListener.QUEUE_DESTINATION;
 import static io.jexxa.infrastructure.utils.messaging.TopicListener.TOPIC_DESTINATION;
@@ -43,7 +46,6 @@ class JMSSenderIT
     private QueueListener queueListener;
     private JexxaMain jexxaMain;
 
-    private MessageSender objectUnderTest;
 
     @BeforeEach
     void initTests()
@@ -51,7 +53,6 @@ class JMSSenderIT
         jexxaMain = new JexxaMain(JexxaTestApplication.class);
         topicListener = new TopicListener();
         queueListener = new QueueListener();
-        objectUnderTest = MessageSenderManager.getMessageSender(JMSSenderIT.class, jexxaMain.getProperties());
 
         jexxaMain.disableBanner()
                 .bind(JMSAdapter.class).to(queueListener)
@@ -59,11 +60,22 @@ class JMSSenderIT
                 .start();
     }
 
-
-    @Test
-    void sendMessageToTopic()
+    @SuppressWarnings("unused")
+    static Stream<Class<? extends MessageSender>> getMessageSenderConfig()
     {
-        //Arrange --
+        return Stream.of(JMSSender.class, TransactionalOutboxSender.class);
+    }
+
+    private static final String MESSAGE_SENDER_CONFIG = "getMessageSenderConfig";
+
+    @ParameterizedTest
+    @MethodSource(MESSAGE_SENDER_CONFIG)
+    void sendMessageToTopic(Class<? extends MessageSender> messageSender)
+    {
+        //Arrange
+        MessageSenderManager.setDefaultStrategy(messageSender);
+        var objectUnderTest = MessageSenderManager.getMessageSender(JMSSenderIT.class, jexxaMain.getProperties());
+
 
         //Act
         objectUnderTest
@@ -81,10 +93,13 @@ class JMSSenderIT
     }
 
 
-    @Test
-    void sendMessageToQueue()
+    @ParameterizedTest
+    @MethodSource(MESSAGE_SENDER_CONFIG)
+    void sendMessageToQueue(Class<? extends MessageSender> messageSender)
     {
-        //Arrange --
+        //Arrange
+        MessageSenderManager.setDefaultStrategy(messageSender);
+        var objectUnderTest = MessageSenderManager.getMessageSender(JMSSenderIT.class, jexxaMain.getProperties());
 
         //Act
         objectUnderTest
@@ -100,10 +115,13 @@ class JMSSenderIT
         assertTimeout(Duration.ofSeconds(1), jexxaMain::stop);
     }
 
-    @Test
-    void sendMessageToQueueAsString()
+    @ParameterizedTest
+    @MethodSource(MESSAGE_SENDER_CONFIG)
+    void sendMessageToQueueAsString(Class<? extends MessageSender> messageSender)
     {
-        //Arrange --
+        //Arrange
+        MessageSenderManager.setDefaultStrategy(messageSender);
+        var objectUnderTest = MessageSenderManager.getMessageSender(JMSSenderIT.class, jexxaMain.getProperties());
 
         //Act
         objectUnderTest
@@ -119,10 +137,57 @@ class JMSSenderIT
         assertTimeout(Duration.ofSeconds(1), jexxaMain::stop);
     }
 
+
+    @ParameterizedTest
+    @MethodSource(MESSAGE_SENDER_CONFIG)
+    void sendByteMessageToTopic(Class<? extends MessageSender> messageSender)
+    {
+        //Arrange
+        MessageSenderManager.setDefaultStrategy(messageSender);
+        var objectUnderTest = MessageSenderManager.getMessageSender(JMSSenderIT.class, jexxaMain.getProperties());
+
+        //Act
+        objectUnderTest
+                .sendByteMessage(message)
+                .toTopic(TOPIC_DESTINATION)
+                .addHeader(TYPE, message.getClass().getSimpleName())
+                .asJson();
+
+        //Assert
+        await().atMost(1, TimeUnit.SECONDS).until(() -> !topicListener.getMessages().isEmpty());
+
+        assertDoesNotThrow(() -> (BytesMessage)topicListener.getMessages().get(0));
+        assertTimeout(Duration.ofSeconds(1), jexxaMain::stop);
+    }
+
+    @ParameterizedTest
+    @MethodSource(MESSAGE_SENDER_CONFIG)
+    void sendByteMessageToQueue(Class<? extends MessageSender> messageSender)
+    {
+        //Arrange
+        MessageSenderManager.setDefaultStrategy(messageSender);
+        var objectUnderTest = MessageSenderManager.getMessageSender(JMSSenderIT.class, jexxaMain.getProperties());
+
+        //Act
+        objectUnderTest
+                .sendByteMessage(message)
+                .toQueue(QUEUE_DESTINATION)
+                .addHeader(TYPE, message.getClass().getSimpleName())
+                .asJson();
+
+        //Assert
+        await().atMost(1, TimeUnit.SECONDS).until(() -> !queueListener.getMessages().isEmpty());
+
+        assertDoesNotThrow(() -> (BytesMessage)queueListener.getMessages().get(0));
+        assertTimeout(Duration.ofSeconds(1), jexxaMain::stop);
+    }
+
     @Test
     void sendMessageReconnectQueue() throws JMSException
     {
-        //Arrange --
+        //Arrange
+        MessageSenderManager.setDefaultStrategy(JMSSender.class); // Reconnect is only meaningful for JMSSender
+        var objectUnderTest = MessageSenderManager.getMessageSender(JMSSenderIT.class, jexxaMain.getProperties());
 
         //Act (simulate an error in between sending two messages
         objectUnderTest
@@ -146,44 +211,6 @@ class JMSSenderIT
     }
 
     @Test
-    void sendByteMessageToTopic()
-    {
-        //Arrange --
-
-        //Act
-        objectUnderTest
-                .sendByteMessage(message)
-                .toTopic(TOPIC_DESTINATION)
-                .addHeader(TYPE, message.getClass().getSimpleName())
-                .asJson();
-
-        //Assert
-        await().atMost(1, TimeUnit.SECONDS).until(() -> !topicListener.getMessages().isEmpty());
-
-        assertDoesNotThrow(() -> (BytesMessage)topicListener.getMessages().get(0));
-        assertTimeout(Duration.ofSeconds(1), jexxaMain::stop);
-    }
-
-    @Test
-    void sendByteMessageToQueue()
-    {
-        //Arrange --
-
-        //Act
-        objectUnderTest
-                .sendByteMessage(message)
-                .toQueue(QUEUE_DESTINATION)
-                .addHeader(TYPE, message.getClass().getSimpleName())
-                .asJson();
-
-        //Assert
-        await().atMost(1, TimeUnit.SECONDS).until(() -> !queueListener.getMessages().isEmpty());
-
-        assertDoesNotThrow(() -> (BytesMessage)queueListener.getMessages().get(0));
-        assertTimeout(Duration.ofSeconds(1), jexxaMain::stop);
-    }
-
-    @Test
     void testPasswordFile()
     {
         //Arrange
@@ -191,8 +218,9 @@ class JMSSenderIT
         properties.putAll(jexxaMain.getProperties());
         properties.remove(JexxaJMSProperties.JNDI_PASSWORD_KEY);
         properties.put(JexxaJMSProperties.JNDI_PASSWORD_FILE, "src/test/resources/secrets/jndiPassword");
+        MessageSenderManager.setDefaultStrategy(JMSSender.class);
 
-        objectUnderTest = MessageSenderManager.getMessageSender(JMSSenderIT.class, properties);
+        var objectUnderTest = MessageSenderManager.getMessageSender(JMSSenderIT.class, properties);
 
         //Act
         objectUnderTest