+++ jexxa-core/src/test/java/io/jexxa/infrastructure/drivenadapterstrategy/persistence/jdbc/JDBCKeyValueRepositoryIT.java
@@ -4,7 +4,6 @@ import io.jexxa.TestConstants;
 import io.jexxa.application.domain.model.JexxaEntity;
 import io.jexxa.application.domain.model.JexxaValueObject;
 import io.jexxa.infrastructure.drivenadapterstrategy.persistence.repository.jdbc.JDBCKeyValueRepository;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
@@ -12,12 +11,12 @@ import org.junit.jupiter.api.parallel.Execution;
 import org.junit.jupiter.api.parallel.ExecutionMode;
 
 import java.io.IOException;
-import java.util.Optional;
 import java.util.Properties;
 
 import static io.jexxa.utils.properties.JexxaCoreProperties.JEXXA_APPLICATION_PROPERTIES;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -48,13 +47,6 @@ class JDBCKeyValueRepositoryIT
                 JexxaEntity::getKey,
                 properties
         );
-    }
-
-    @AfterEach
-    void tearDown()
-    {
-        Optional.ofNullable(objectUnderTest)
-                .ifPresent(JDBCKeyValueRepository::close);
     }
 
 
@@ -127,6 +119,57 @@ class JDBCKeyValueRepositoryIT
 
         objectUnderTest.getConnection().close();
         assertDoesNotThrow(this::addAggregate);
+    }
+
+    @Test
+    void testSuccessfulTransaction()
+    {
+        //Arrange
+        objectUnderTest.initTransaction();
+        objectUnderTest.add(aggregate);
+
+        //act
+        objectUnderTest.closeTransaction();
+
+
+        //Assert
+        assertFalse(objectUnderTest.get(aggregate.getKey()).isEmpty());
+    }
+
+    @Test
+    void testTransactionRollback()
+    {
+        //Arrange
+        objectUnderTest.initTransaction();
+        objectUnderTest.add(aggregate);
+
+        //act
+        objectUnderTest.rollback();
+        objectUnderTest.closeTransaction();
+
+
+        //Assert
+        assertTrue(objectUnderTest.get(aggregate.getKey()).isEmpty());
+    }
+
+    @Test
+    void testFailedTransaction()
+    {
+        //Arrange
+        objectUnderTest.initTransaction();
+        objectUnderTest.add(aggregate);
+
+        //act
+        try {
+            objectUnderTest.add(aggregate);
+        } catch (RuntimeException e) {
+            objectUnderTest.rollback();
+        }
+        objectUnderTest.closeTransaction();
+
+
+        //Assert
+        assertTrue(objectUnderTest.get(aggregate.getKey()).isEmpty());
     }
 
 }